+++ WinPort/src/sudo/sudo_private.h
@@ -43,7 +43,7 @@ namespace Sudo
 	{
 		LocalSocket &_sock;
 		bool _failed;
-		
+
 	public:
 		BaseTransaction(LocalSocket &sock);
 		void SendBuf(const void *buf, size_t len);
@@ -59,30 +59,30 @@ namespace Sudo
 		int RecvInt();
 		inline int RecvFD() { return _sock.RecvFD(); }
 		inline void RecvErrno() { errno = RecvInt(); }
-		
+
 		inline bool IsFailed() const { return _failed; }
 	};
 
 	class ClientTransaction : protected std::lock_guard<std::mutex>, public BaseTransaction
 	{
 		SudoCommand _cmd;
-		
+
 		void Finalize();
 		public:
 		ClientTransaction(SudoCommand cmd);
 		~ClientTransaction();
-		
+
 		void NewTransaction(SudoCommand cmd);
 	};
 
 	bool TouchClientConnection(bool want_modify);
 	bool IsSudoRegionActive();
-	
+
 	void ClientCurDirOverrideReset();
 	void ClientCurDirOverrideSet(const char *path);
 	bool ClientCurDirOverrideSetIfRecent(const char *path);
 	bool ClientCurDirOverrideQuery(char *path, size_t size);
-	
+
 	class ClientReconstructCurDir
 	{
 		char *_free_ptr;
@@ -94,7 +94,7 @@ namespace Sudo
 	};
 
 #if !defined(__APPLE__) && !defined(__FreeBSD__)
-	int bugaware_ioctl_pint(int fd, unsigned long req, int *v);
+	int bugaware_ioctl_pint(int fd, unsigned long req, unsigned long *v);
 #endif
 
 }
